Crate problem070

Problem 70 (Totient permutation)

Solution detail

The product formula for ϕ(n) tells us that n / ϕ(n) is equal to the product, over all primes p dividing n, of the quantity p / (p - 1).

This quantity is always greater than 1, and is decreasing as p gets larger, so n / ϕ(n) is minimised when n is the product of a small number of large primes - that is, when n is itself prime, or when n is a square of a prime, or a semiprime composed of two close-together primes...

It is easy to check that when n is prime, ϕ(n) = n - 1 is never a permutation of n, so we should restrict our search to n the product of two (not necessarily distinct) primes. This can be done quickly by:

Pre-sieving the primes up to n / 2

Checking, for n = p * q for each pair p, q, whether ϕ(n) is a permutation of n

Keeping track of the smallest ratio n / ϕ(n) for solutions as they are found.

Functions

answer
Solve the problem, returning the answer as a String

are_permutations
Determine whether the two given numbers are permutations of one another.

main

solve
Find the value of n not exceeding the given limit for which ϕ(n) is a permutation of n and the ratio n / ϕ(n) is minimised.